LeafIQ/requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
dm-tree==0.1.8
filelock==3.13.1
flatbuffers==23.5.26
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.0
h5py==3.10.0
idna==3.6
keras==3.0.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
protobuf==4.25.3
Pygments==2.17.2
pytz==2024.1
requests==2.31.0
rich==13.7.0
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.0rc0
tensorflow-intel==2.16.0rc0
termcolor==2.4.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.16.0
